+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog].
 - Night Config Fixes now also handles individual config values being corrected: 
   - In Forge for those cases always the built-in default value for these options is used if the current value cannot be read
   - Now the default config in `defaultconfigs` is checked first if it contains an entry for the value being corrected
+- `ModConfig` is no longer wrapped internally, changes to final fields are now applied using `Unsafe`
 
 ## [v4.0.2-1.19.2] - 2023-02-03
 ### Changed